Fixing PEP8

+++ math/0x02-calculus/9-sum_total.py
@@ -11,5 +11,5 @@ def summation_i_squared(n):
     """a function that calculates \sum_{i=1}^{n} i^2"""
     if type(n) != int or n < 1:
         return None
-    result = int((n * (n + 1) * (2 * n + 1)) / 6)
+    result = n * (n + 1) * (2 * n + 1)) / 6
     return result
